refactor(lld): give the emulated screen image a view model
