+++ apps/ledger-live-desktop/src/mvvm/components/DeviceScreen/components/DeviceScreenImage/__tests__/DeviceScreenImage.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { fireEvent, render, screen } from "tests/testSetup";
-import { DeviceScreenImage } from "../DeviceScreenImage";
+import { DeviceScreenImage } from "../index";
 
 /**
  * jsdom ships no PointerEvent, and the fallback it leaves carries no
