+++ apps/ledger-live-desktop/src/mvvm/components/DeviceScreen/components/DeviceScreenImage/index.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { DeviceScreenImageView } from "./DeviceScreenImageView";
+import {
+  useDeviceScreenImageViewModel,
+  type DeviceScreenTouch,
+} from "./useDeviceScreenImageViewModel";
+
+export interface DeviceScreenImageProps {
+  readonly src: string;
+  /** Omitted on button-driven devices, which are not tappable. */
+  readonly onTouch?: DeviceScreenTouch;
+}
+
+export function DeviceScreenImage({ src, onTouch }: DeviceScreenImageProps) {
+  return <DeviceScreenImageView src={src} {...useDeviceScreenImageViewModel(src, onTouch)} />;
+}
